French keyboard improvement
Hi,
I'm using SHR for several weeks, and the current
illume-keyboard-french doesn't satisfy me...
So, I've made some modifications that may interest other people as well.

First, I've inverted normal and shift keys for the first raw.
By doing this, this layout looks like a lot more to the latin9 layout.
This choice might be disputable, but IMHO, it's better like that.

Second, I've moved '['/'{' and ']'/'}' to the bottom of the keyboard.
Again, this choice may not be the best, but I don't think it's doing any
harm.

Eventually, the previous change, altogether with the regroupement of the
'#' and '%' (which have moved to where it is in latin9) keys, permits
the '^' and '¨' dead keys, needed to type common characters in the
French language like 'î', 'â', 'ë', 'ï' or 'ô'.

Best Regards,
Thibaut GIRKA.